Refactoring

+++ src/grib_iterator_class_gaussian_reduced.c
@@ -146,7 +146,7 @@ static int iterate_reduced_gaussian_subarea(grib_iterator* iter, grib_handle* h,
 
     if (h->context->debug) {
         const size_t np = count_subarea_points(h, get_reduced_row, pl, plsize, lon_first, lon_last);
-        printf("ECCODES DEBUG grib_iterator_class_gaussian_reduced: sub-area num points=%ld\n", np);
+        printf("ECCODES DEBUG grib_iterator_class_gaussian_reduced: sub-area num points=%ld\n", (long)np);
     }
 
     /*find starting latitude */
@@ -282,7 +282,7 @@ static int init(grib_iterator* iter, grib_handle* h, grib_arguments* args)
         iter->e=0;
         if (h->context->debug) {
             const size_t np = count_global_points(pl, plsize);
-            printf("ECCODES DEBUG grib_iterator_class_gaussian_reduced: global num points=%ld\n", np);
+            printf("ECCODES DEBUG grib_iterator_class_gaussian_reduced: global num points=%ld\n", (long)np);
         }
 
         for (j=0;j<plsize;j++) {
